Extract file buildSubscriptionsState
