+++ prelaunch/list-subscribers/endpoint.js
@@ -3,6 +3,7 @@ var async = require("async");
 var fs = require("fs");
 var path = require("path");
 var renderView = require("../../renderView.js");
+var buildSubscriptionsState = require("./buildSubscriptionsState.js");
 
 function listSubscribersEndpoint(pool, req, res) {
     if(req.query.secret != "1241215124125123122215124") {
@@ -29,40 +30,7 @@ function listSubscribersEndpoint(pool, req, res) {
 }
 
 function getActiveSubscriptions(pool, callback) {
-    pool.query("SELECT * FROM user_signed_up_for_newsletter", (error, result) => {
-        if(error) {
-            console.error("[getActiveSubscriptions] Failed to get subscription list", error);
-            return callback(error);
-        }
-        var subscriptionContents = result.rows.map(row => {
-            return {
-                id: row.id,
-                email: row.data.email
-            };
-        });
-        pool.query("SELECT * FROM user_unsubscribed_from_newsletter", (error, result) => {
-            if(error) {
-                console.error("[getActiveSubscriptions] Failed to check who has unsubscribed for subscriber list", error);
-                return callback(error);
-            }
-            var unsubscribedEmails = result.rows.map(row => {
-                var match = subscriptionContents.filter(subscription => subscription.id == row.id)[0];
-                return match ? match.email : null;
-            });
-            var alreadyIncludedEmails = [];
-            subscriptionContents = subscriptionContents.filter((subscription, index) => {
-                if(unsubscribedEmails.indexOf(subscription.email) != -1) {
-                    return false;
-                }
-                if(alreadyIncludedEmails.indexOf(subscription.email) != -1) {
-                    return false;
-                }
-                alreadyIncludedEmails.push(subscription.email);
-                return true;
-            });
-            callback(null, subscriptionContents);
-        });
-    });
+    buildSubscriptionsState(pool, callback);
 }
 
 module.exports = function(pool) {
